+++ HISTORY.rst
@@ -4,6 +4,19 @@ History
 =======
 
 
+0.6.0b0 (2016-05-13)
+--------------------
+* Small debug output reduction
+* Not running warning
+
+
+0.6.0a0 (2016-05-13)
+--------------------
+* Change cli
+* Change check/check_all interfaces (Use check_all like check before)
+* Make StartStopable
+
+
 0.5.4a0 (2016-05-05)
 --------------------
 * First release